#1092 Let no file end the dive session, whatever it raises
The guard caught five named types. Which one a damaged file raises is decided by
the decoder that trips over it — a dictionary index past its dictionary is an
ArrayIndexOutOfBoundsException, an impossible RLE run header an
IllegalStateException, a length that will not fit an ArithmeticException, a
declared size the buffer cannot hold a NegativeArraySizeException — and the four
regions these screens parse go at the bytes outside the read pipeline, so
nothing narrows that set before the guard sees it. A list has to be revisited
every time a decoder learns a new way to fail, and the one thing the guard
exists to promise is that no file ends the session.

Two keys also outlived their screen. `o` collapses to Overview without clearing
the failure, so the overlay stayed painted over a screen that reads perfectly
well; Esc could not dismiss it there, because both back-navigation arms want a
frame to pop and Overview ignores Esc itself. A reader who pressed `o` was left
with an error box and no key but `q`.

The chunk length `inspect pages` reads a column chunk with was narrowed with a
cast, which turns a corrupt total_compressed_size into a wrong read rather than
a raised one.

+++ cli/src/main/java/dev/hardwood/cli/command/InspectPagesCommand.java
@@ -443,7 +443,7 @@ public class InspectPagesCommand implements Command<CommandInvocation> {
         long chunkStart = (dictOffset != null && dictOffset > 0) ? dictOffset : cmd.dataPageOffset();
         long chunkSize = cmd.totalCompressedSize();
 
-        ByteBuffer buffer = inputFile.readRange(chunkStart, (int) chunkSize);
+        ByteBuffer buffer = inputFile.readRange(chunkStart, Math.toIntExact(chunkSize));
 
         List<PageInfo> rows = new ArrayList<>();
         int pageIndex = 0;
